requirements.txt
Django==1.11
Pillow==5.0.0
django-tastypie==0.14.0